Set up infrastructure to get PDFs from ADS
This turns out to be a bit tricky since ADS will return a PDF URL even if it
doesn't actually have a PDF -- that URL will just 404 when we download it.
Catch that condition.

Also add code to check that the fetched "PDF" is in fact a PDF document, at
least according to its first four bytes.