Clean up password-storage situation
As described in the comments, I'm kinda-sorta copying what Firefox does
for website passwords. This is mainly in the vague sense that we need to
store the plaintext passwords, and we don't require any user input to
retrieve them, so we're inherently insecure. We do some nontrivial
encryption to prevent accidental disclosure, but it's not worth trying
harder because anyone with access to the files wins.