Preliminary work on BibTeX merging.
I realized that my `latexdriver` approach to managing Bibtex files has a
serious flaw: you can't compile anything unless you have a `bibtools` database
fully populated with information for all of the referenced publications, and
of course that completely breaks if I'm ever writing a document
collaboratively with anyone since they're not going to have the same database
as me.

On the other hand, automatically filling the .bib file from my database is
really cool and convenient. And I want to maintain the existing status quo
where [1] the bibtex file is kept in version control (since it has to be for
an arbitrary third part to be able to compile the document) and [2] the bibtex
file is not a grab-bag of every reference I've ever read, but instead contains
only the necessary information.

So, we are going to need some infrastructure to parse Bibtex files and merge
records. This is still not fully baked, but here's a first cut with some of
the necessary features. Namely, the Bibtex output routines can now output a
record that doesn't have a corresponding entry in the database.

Again, there's much that I still need to think through here, but this is at
least a start.